requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
pydantic==2.9.2
python-multipart==0.0.12
Pillow==10.4.0
numpy==1.26.4
ultralytics==8.3.10
